+++ src/main/java/com/example/labschedulerserver/repository/CourseRepository.java
@@ -8,11 +8,11 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface CourseRepository extends JpaRepository<Course, Integer> {
+public interface CourseRepository extends JpaRepository<Course, Long> {
 //    @Query(value = "SELECT EXISTS (SELECT 1 FROM Course WHERE subject_id = ?1 AND class_id = ?2 AND semester_id = ?3)", nativeQuery = true)
 //    public boolean checkCourseExist(Integer subjectId, Integer classId, Integer semesterId);
 
-    public Course findCoursesBySubjectIdAndClazzIdAndSemesterId(Integer subjectId, Integer classId, Integer semesterId);
+    public Course findCoursesBySubjectIdAndClazzIdAndSemesterId(Long subjectId, Long classId, Long semesterId);
 
-    public List<Course> findAllBySemesterId(Integer semesterId);
+    public List<Course> findAllBySemesterId(Long semesterId);
 }
