+++ src/main/java/com/example/labschedulerserver/repository/MajorRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MajorRepository extends JpaRepository<Major, Integer> {
+public interface MajorRepository extends JpaRepository<Major, Long> {
     public Major findMajorByName(String name);
 }
