+++ src/main/java/com/example/labschedulerserver/repository/StudentAccountRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StudentAccountRepository extends JpaRepository<StudentAccount, Integer> {
+public interface StudentAccountRepository extends JpaRepository<StudentAccount, Long> {
     public Object findByAccount(Account account);
 }
