+++ src/main/java/com/example/labschedulerserver/repository/SubjectRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface SubjectRepository  extends JpaRepository<Subject, Integer> {
+public interface SubjectRepository  extends JpaRepository<Subject, Long> {
     Optional<Subject> findByName(String name);
 }
