+++ src/main/java/com/example/labschedulerserver/service/RoomService.java
@@ -8,9 +8,9 @@ import java.util.Map;
 
 public interface RoomService {
     public List<Room> getAllRoom();
-    public Room getRoomById(Integer id);
+    public Room getRoomById(Long id);
     public Room addNewRoom(AddRoomRequest addRoomRequest);
-    public void deleteRoomById(Integer id);
-    public Room updateRoomById(Integer id, Map<String,Object> payload);
+    public void deleteRoomById(Long id);
+    public Room updateRoomById(Long id, Map<String,Object> payload);
     public Room findByName(String name);
 }
